gl_shader_gen: use HashableStruct for config structs
To avoid padding issue. The HashableStruct requires that the underlying struct cannot have a user-defined ctor (because they would be overridden by memset), so changed all existing ctor to Init functions, and let the derived type ctor calls Init.

The std::pair in PicaGSConfigCommon(Raw) is changed to an explicit struct because std::pair has a user-defined ctor (value-initialization).